tools/crawlers/requirements.txt
gspread>=6.0
google-auth>=2.0
google-api-python-client>=2.0
beautifulsoup4>=4.12
requests>=2.31
